+++ CHANGELOG.md
@@ -5,6 +5,18 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 [Keep a CHANGELOG](http://keepachangelog.com).
 
 
+## Unreleased
+
+### Added
+
+- `RDF.Triple.valid?/1`, `RDF.Quad.valid?/1` and `RDF.Statement.valid?/1`, which
+  validate if a tuple is a valid RDF triple or RDF quad
+
+
+[Compare v0.5.2...HEAD](https://github.com/marcelotto/rdf-ex/compare/v0.5.2...HEAD)
+
+
+
 ## 0.5.2 - 2018-11-04
 
 ### Added